+++ gridcapa-cse-valid-app/src/main/java/com/farao_community/farao/cse_valid/app/dichotomy/DichotomyRunner.java
@@ -6,7 +6,9 @@
  */
 package com.farao_community.farao.cse_valid.app.dichotomy;
 
+import com.farao_community.farao.cse_valid.api.exception.CseValidInvalidDataException;
 import com.farao_community.farao.cse_valid.api.resource.CseValidRequest;
+import com.farao_community.farao.cse_valid.app.CseValidHandler;
 import com.farao_community.farao.cse_valid.app.FileImporter;
 import com.farao_community.farao.dichotomy.api.DichotomyEngine;
 import com.farao_community.farao.dichotomy.api.NetworkShifter;
@@ -15,10 +17,13 @@ import com.farao_community.farao.dichotomy.api.index.RangeDivisionIndexStrategy;
 import com.farao_community.farao.dichotomy.api.results.DichotomyResult;
 import com.farao_community.farao.dichotomy.shift.LinearScaler;
 import com.farao_community.farao.dichotomy.shift.SplittingFactors;
+import com.farao_community.farao.minio_adapter.starter.MinioAdapter;
 import com.farao_community.farao.rao_runner.api.resource.RaoResponse;
 import com.powsybl.glsk.api.GlskDocument;
 import com.powsybl.iidm.network.Network;
 import com.rte_france.farao.cep_seventy_validation.timestamp_validation.ttc_adjustment.TTimestamp;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
 
 import java.io.IOException;
@@ -29,16 +34,19 @@ import java.io.IOException;
 @Service
 public class DichotomyRunner {
 
+    private static final Logger LOGGER = LoggerFactory.getLogger(CseValidHandler.class);
     private static final RangeDivisionIndexStrategy INDEX_STRATEGY_CONFIGURATION = new RangeDivisionIndexStrategy(false);
     private final FileImporter fileImporter;
+    private final MinioAdapter minioAdapter;
     private GlskDocument glskDocument;
     private Network network;
 
-    public DichotomyRunner(FileImporter fileImporter) {
+    public DichotomyRunner(FileImporter fileImporter, MinioAdapter minioAdapter) {
         this.fileImporter = fileImporter;
+        this.minioAdapter = minioAdapter;
     }
 
-    public DichotomyResult<RaoResponse> runDichotomy(CseValidRequest cseValidRequest, TTimestamp timestamp) throws IOException {
+    public DichotomyResult<RaoResponse> runDichotomy(CseValidRequest cseValidRequest, TTimestamp timestamp) {
         importFiles(cseValidRequest);
         int npAugmented = timestamp.getMNII().getV().intValue();
         int np = timestamp.getMiBNII().getV().intValue() - timestamp.getANTCFinal().getV().intValue();
@@ -57,12 +65,29 @@ public class DichotomyRunner {
     }
 
     private void importFiles(CseValidRequest cseValidRequest) {
-
+        try {
+            this.glskDocument = importGlskFile(cseValidRequest);
+            this.network = importNetworkFile(cseValidRequest);
+        } catch (IOException e) {
+            LOGGER.error("Can not import files");
+            throw new CseValidInvalidDataException("Can not import files");
+        }
     }
 
-    private NetworkShifter getNetworkShifter() throws IOException {
-        return new LinearScaler(
-                fileImporter.importGlsk("").getZonalScalable(null),
+    private NetworkShifter getNetworkShifter() {
+        return new LinearScaler(glskDocument.getZonalScalable(network),
                 new SplittingFactors(null));
+    }
+
+    public GlskDocument importGlskFile(CseValidRequest cseValidRequest) throws IOException {
+        String url = fileImporter.buildGlskFileUrl(cseValidRequest);
+        String file = minioAdapter.generatePreSignedUrl(url);
+        return fileImporter.importGlsk(file);
+    }
+
+    public Network importNetworkFile(CseValidRequest cseValidRequest) throws IOException {
+        String url = fileImporter.buildNetworkFileUrl(cseValidRequest);
+        String fileUrl = minioAdapter.generatePreSignedUrl(url);
+        return fileImporter.importNetwork(cseValidRequest.getCgm().getFilename(), fileUrl);
     }
 }
